Remove PrngState of some functions
Although random streams needs a PRNG and a seed to flow,
it is too heavy to keep the PRNG in the flow beacause the PRNG is invariable.
Therefore, it seems good for random stream to get a PRNG earlier than others,
and PrngState is no longer necessary in the workflow.